+++ src/main/resources/application.properties
@@ -5,4 +5,4 @@ spring.datasource.url= jdbc:postgresql://localhost:5432/postgres
 spring.datasource.username=postgres
 spring.datasource.password=123456
 spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=create-drop
